requirements.txt
torch
numpy
torchvision
Pillow
PyYAML